refactor(server): move readiness probes into service
